% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allGOterms}
\alias{allGOterms}
\title{allGOterms: data.frame with ids and terms}
\format{data.frame instance}
\source{
GO.db, August 2017
}
\usage{
allGOterms
}
\description{
allGOterms: data.frame with ids and terms
}
\keyword{datasets}
