% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdf.R
\name{buildCellOntSupport}
\alias{buildCellOntSupport}
\title{read and model the Cell Ontology as shipped in OWL and parsed in redland}
\usage{
buildCellOntSupport()
}
\value{
instance of \code{\link{RrdfSupport-class}}
}
\description{
read and model the Cell Ontology as shipped in OWL and parsed in redland
}
\examples{
buildCellOntSupport()
}
