% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdf.R
\name{buildEFOOntSupport}
\alias{buildEFOOntSupport}
\title{read and model the Experimental Factor Ontology as shipped in OWL and parsed in redland}
\usage{
buildEFOOntSupport()
}
\value{
instance of \code{\link{RrdfSupport-class}}
}
\description{
read and model the Experimental Factor Ontology as shipped in OWL and parsed in redland
}
\examples{
buildEFOOntSupport()
}
