% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\name{c,TermSet-method}
\alias{c,TermSet-method}
\title{combine TermSet instances}
\usage{
\S4method{c}{TermSet}(x, ...)
}
\arguments{
\item{x}{TermSet instance}

\item{\dots}{additional instances}
}
\value{
TermSet instance
}
\description{
combine TermSet instances
}
