% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\name{cellTypeToGO}
\alias{cellTypeToGO}
\alias{cellTypeToGenes}
\title{utilities for approximate matching of cell type terms to GO categories and annotations}
\usage{
cellTypeToGO(celltypeString, ...)

cellTypeToGenes(celltypeString, orgDb, cols = c("ENSEMBL", "SYMBOL"), ...)
}
\arguments{
\item{celltypeString}{character atom to be used to search GO terms using 
\code{\link[base]{agrep}}}

\item{\dots}{additional arguments to \code{\link[base]{agrep}}}
}
\description{
utilities for approximate matching of cell type terms to GO categories and annotations
}
