% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clfixer.R
\name{cleanCLOnames}
\alias{cleanCLOnames}
\title{obtain named character vector of terms from Cell Line Ontology, omitting obsolete and trailing 'cell'}
\usage{
cleanCLOnames()
}
\value{
character()
}
\description{
obtain named character vector of terms from Cell Line Ontology, omitting obsolete and trailing 'cell'
}
\examples{
cleanCLOnames()[1:10]
}
