% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLfeats.R
\name{CLfeats}
\alias{CLfeats}
\title{produce a data.frame of features relevant to a Cell Ontology class}
\usage{
CLfeats(ont, tag = "CL:0001054")
}
\arguments{
\item{ont}{instance of ontologyIndex ontology}

\item{tag}{character(1) a CL: class tag}
}
\value{
a data.frame instance
}
\description{
produce a data.frame of features relevant to a Cell Ontology class
}
\note{
This function will look in the intersection_of and has_part,
lacks_part components of the CL entry to find properties asserted
of or inherited by the cell type identified in 'tag'
}
\examples{
cl = getCellOnto()
pr = getPROnto()
go = getGeneOnto()
CLfeats(cl, tag="CL:0001054")
}
