% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_classes.R
\name{connect_classes}
\alias{connect_classes}
\title{connect ontological categories between related, annotated SummarizedExperiments}
\usage{
connect_classes(ont, se1, se2)
}
\arguments{
\item{ont}{an ontologyIndex ontology instance}

\item{se1}{SummarizedExperiment instance with `label.ont` among colData columns}

\item{se2}{SummarizedExperiment instance with `label.ont` among colData columns}
}
\value{
a list with two sublists mapping from terms in one SE to descendant terms in the other SE
}
\description{
connect ontological categories between related, annotated SummarizedExperiments
}
