% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmarks.R
\name{ctmarks}
\alias{ctmarks}
\title{app to review molecular properties of cell types via cell ontology}
\usage{
ctmarks()
}
\description{
app to review molecular properties of cell types via cell ontology
}
\note{
Prototype of harvesting of cell ontology by searching
has_part, has_plasma_membrane_part, intersection_of and allied
ontology relationships.  Uses shiny.
}
