% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{demoApp}
\alias{demoApp}
\title{demonstrate the use of makeSelectInput}
\usage{
demoApp()
}
\value{
Run only for side effect of starting a shiny app.
}
\description{
demonstrate the use of makeSelectInput
}
\examples{
  if (interactive()) {
require(shiny)
print(demoApp())
}
}
