% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropStop.R
\name{dropStop}
\alias{dropStop}
\title{dropStop is a utility for removing certain words from text data}
\usage{
dropStop(x, drop, lower = TRUE, splitby = " ")
}
\arguments{
\item{x}{character vector of strings to be cleaned}

\item{drop}{character vector of words to scrub}

\item{lower}{logical, if TRUE, x converted with \code{\link{tolower}}}

\item{splitby}{character, used with strsplit to tokenize \code{x}}
}
\value{
a list with one element per input string, split by " ", with elements in \code{drop} removed
}
\description{
dropStop is a utility for removing certain words from text data
}
\examples{
data(minicorpus)
minicorpus[1:3]
dropStop(minicorpus)[1:3]
}
