% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{getCellOnto}
\alias{getCellOnto}
\alias{getCellLineOnto}
\alias{getEFOOnto}
\alias{getChebiLite}
\title{load ontologies that may include non-ascii strings and therefore cannot be in data folder}
\usage{
getCellOnto()

getCellLineOnto()

getEFOOnto()

getChebiLite()
}
\description{
load ontologies that may include non-ascii strings and therefore cannot be in data folder
}
\examples{
co = getCellOnto()
co
clo = getCellLineOnto()
length(clo$id)
che = getChebiLite()
length(che$id)
efo = getEFOOnto()
length(efo$id)
}
