% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeproc.R
\name{getLeavesFromTerm}
\alias{getLeavesFromTerm}
\title{obtain childless descendents of a term (including query)}
\usage{
getLeavesFromTerm(x, ont)
}
\arguments{
\item{x}{a character(1) id element for ontology_index instance}

\item{ont}{an ontology_index instance as defined in ontologyIndex package}
}
\value{
character vector of 'leaves' of ontology tree
}
\description{
obtain childless descendents of a term (including query)
}
\examples{
ch = getChebiOnto()
alldr = getLeavesFromTerm("CHEBI:23888", ch)
head(ch$name[alldr[1:15]])
}
