% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdf.R
\name{getModel}
\alias{getModel}
\alias{getWorld}
\title{accessors for RrdfSupport}
\usage{
getModel(x)

getWorld(x)
}
\arguments{
\item{x}{instance of RrdfSupport}
}
\value{
instance of \code{\link[redland]{Model-class}} from redland

instance of \code{\link[redland]{World-class}} from redland
}
\description{
accessors for RrdfSupport
}
