% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humrna}
\alias{humrna}
\title{humrna: a data.frame of SRA metadata related to RNA-seq in humans}
\format{
data.frame
}
\source{
NCBI SRA
}
\usage{
humrna
}
\description{
humrna: a data.frame of SRA metadata related to RNA-seq in humans
}
\note{
arbitrarily chosen from RNA-seq studies for taxon 9606
}
\examples{
data(humrna)
names(humrna)
head(humrna[,1:5])
}
\keyword{datasets}
