% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL.R
\name{improveNodes}
\alias{improveNodes}
\title{inject linefeeds for node names for graph, with textual
annotation from ontology}
\usage{
improveNodes(g, ont)
}
\arguments{
\item{g}{graphNEL instance}

\item{ont}{instance of ontology from ontologyIndex}
}
\description{
inject linefeeds for node names for graph, with textual
annotation from ontology
}
