% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL.R
\name{make_graphNEL_from_ontology_plot}
\alias{make_graphNEL_from_ontology_plot}
\title{obtain graphNEL from ontology_plot instance of ontologyPlot}
\usage{
make_graphNEL_from_ontology_plot(x)
}
\arguments{
\item{x}{instance of S3 class ontology_plot}
}
\value{
instance of S4 graphNEL class
}
\description{
obtain graphNEL from ontology_plot instance of ontologyPlot
}
\examples{
requireNamespace("Rgraphviz")
requireNamespace("graph")
cl = getCellOnto()
cl3k = c("CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
   "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556")
p3k = ontologyPlot::onto_plot(cl, cl3k)
gnel = make_graphNEL_from_ontology_plot(p3k)
gnel = improveNodes(gnel, cl)
graph::graph.par(list(nodes=list(shape="plaintext", cex=.8)))
gnel = Rgraphviz::layoutGraph(gnel)
Rgraphviz::renderGraph(gnel)
}
