% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{makeSelectInput}
\alias{makeSelectInput}
\title{generate a selectInput control for an ontologyIndex slice}
\usage{
makeSelectInput(
  onto,
  term,
  type = "siblings",
  inputId,
  label,
  multiple = TRUE,
  ...
)
}
\arguments{
\item{onto}{ontologyIndex instance}

\item{term}{character(1) term used as basis for term list option set in the control}

\item{type}{character(1) 'siblings' or 'children', relationship to 'term' that the options will satisfy}

\item{inputId}{character(1) for use in server}

\item{label}{character(1) for labeling in ui}

\item{multiple}{logical(1) passed to \code{\link[shiny]{selectInput}}}

\item{\dots}{additional parameters passed to \code{\link[shiny]{selectInput}}}
}
\value{
a \code{\link[shiny]{selectInput}} control
}
\description{
generate a selectInput control for an ontologyIndex slice
}
\examples{
makeSelectInput
}
