% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_classes.R
\name{map2prose}
\alias{map2prose}
\title{use prose terminology with output of connect_classes}
\usage{
map2prose(x, cl)
}
\arguments{
\item{x}{a component of connect_classes output}

\item{cl}{an ontologyIndex ontology instance}
}
\value{
a decorated list
}
\description{
use prose terminology with output of connect_classes
}
