% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNaive.R
\name{mapOneNaive}
\alias{mapOneNaive}
\title{use grep or agrep to find a match for a naive token into ontology}
\usage{
mapOneNaive(naive, onto, useAgrep = FALSE, ...)
}
\arguments{
\item{naive}{character(1)}

\item{onto}{an instance of ontologyIndex::ontology_index}

\item{useAgrep}{logical(1) if TRUE, agrep will be used}

\item{\dots}{passed to agrep if used}
}
\value{
if a match is found, the result of grep/agrep with value=TRUE is returned; otherwise a named NA_character_ is returned

named vector, names are ontology identifiers, values are matched strings
}
\description{
use grep or agrep to find a match for a naive token into ontology
}
\examples{
co = ontoProc::getCellOnto()
mapOneNaive("astrocyte", co)
}
