% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minicorpus}
\alias{minicorpus}
\title{minicorpus: a vector of annotation strings found in 'study title' of SRA metadata.}
\format{
character vector
}
\source{
NCBI SRA
}
\usage{
minicorpus
}
\description{
minicorpus: a vector of annotation strings found in 'study title' of SRA metadata.
}
\note{
arbitrarily chosen from titles of RNA-seq studies for taxon 9606
}
\examples{
data(minicorpus)
head(minicorpus)
}
\keyword{datasets}
