% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{onto_roots}
\alias{onto_roots}
\title{list parentless nodes in ontology_index instance}
\usage{
onto_roots(x)
}
\arguments{
\item{x}{an ontology_index instance}
}
\value{
a report (produced by cat()) of root ids and associated names
}
\description{
list parentless nodes in ontology_index instance
}
\examples{
onto_roots
}
