% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{packDesc2019}
\alias{packDesc2019}
\title{packDesc2019: overview of ontoProc resources}
\format{
data.frame instance
}
\usage{
packDesc2019
}
\description{
packDesc2019: overview of ontoProc resources
}
\note{
Brief survey of functions available to load serialized
ontology_index instances imported from OBO.
}
\examples{
head(packDesc2019)
}
\keyword{datasets}
