% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROSYM}
\alias{PROSYM}
\title{PROSYM: HGNC symbol synonyms for PR (protein ontology) entries identified in Cell Ontology}
\format{
data.frame instance
}
\source{
OBO Foundry
}
\usage{
PROSYM
}
\description{
PROSYM: HGNC symbol synonyms for PR (protein ontology) entries identified in Cell Ontology
}
\note{
This is a snapshot of the synonyms
component of an extract_tags='everything' import of PR.
The 'EXACT.*PRO-short.*:DNx' pattern is used to retrieve
HGNC symbols.
See ?getPROnto for more provenance information.
}
\examples{
head(ontoProc::PROSYM)
}
\keyword{datasets}
