% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLfeats.R
\name{recognizedPredicates}
\alias{recognizedPredicates}
\title{enumerate ontological relationships used in ontoProc utilities}
\usage{
recognizedPredicates()
}
\value{
character vector, names of elements are abbreviated tokens that may be used in code
}
\description{
enumerate ontological relationships used in ontoProc utilities
}
\examples{
head(recognizedPredicates())
}
