% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdf.R
\docType{class}
\name{RrdfSupport-class}
\alias{RrdfSupport-class}
\alias{show,RrdfSupport-method}
\title{container for RDF model and world instances as defined in redland}
\usage{
\S4method{show}{RrdfSupport}(object)
}
\arguments{
\item{object}{instance of RrdfSupport}
}
\description{
container for RDF model and world instances as defined in redland
}
