% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\name{secLevGen}
\alias{secLevGen}
\title{simple generation of children of 'choices' given as terms, returned as TermSet}
\usage{
secLevGen(choices, ont)
}
\arguments{
\item{choices}{vector of terms}

\item{ont}{instance of ontology_index (S3) from ontologyIndex package}
}
\value{
TermSet instance
}
\description{
simple generation of children of 'choices' given as terms, returned as TermSet
}
\examples{
efoOnto = getEFOOnto()
secLevGen( "disease", efoOnto )
}
