% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurTab.R
\name{seur3kTab}
\alias{seur3kTab}
\title{tabulate the basic outcome of PBMC 3K tutorial of Seurat}
\usage{
seur3kTab()
}
\value{
a data.frame
}
\description{
tabulate the basic outcome of PBMC 3K tutorial of Seurat
}
\examples{
seur3kTab()
}
