% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\name{siblings_URL}
\alias{siblings_URL}
\alias{label_URL}
\alias{children_URL}
\title{generate a TermSet with siblings of a given term}
\usage{
siblings_URL(urlstring = "<http://www.ebi.ac.uk/efo/EFO_1001209>", model,
  world)

label_URL(urlstring = "<http://www.ebi.ac.uk/efo/EFO_0000311>", model, world)

children_URL(urlstring = "<http://www.ebi.ac.uk/efo/EFO_0000787>", model,
  world)
}
\arguments{
\item{urlstring}{a URI that serves as an rdf-schema class}

\item{model}{RDF model instance as defined in redland package}

\item{world}{RDF world instance as defined in redland package}
}
\description{
generate a TermSet with siblings of a given term

acquire the label of an ontology subject tag

acquire the label of an ontology subject tag
}
\examples{
if (!exists(".efosupp")) .efosupp = buildEFOOntSupport()
siblings_URL( model=getModel(.efosupp), world=getWorld(.efosupp))
}
