% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopWords}
\alias{stopWords}
\title{stopWords: vector of stop words from xpo6.com}
\format{
character vector
}
\source{
\url{http://xpo6.com/list-of-english-stop-words/}
}
\usage{
stopWords
}
\description{
stopWords: vector of stop words from xpo6.com
}
\note{
"Stop words" are english words that are assumed to contribute limited 
semantic value in the analysis of free text.
}
\examples{
data(stopWords)
head(stopWords)
}
\keyword{datasets}
