% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allGOterms}
\alias{allGOterms}
\title{allGOterms: data.frame with ids and terms}
\format{
data.frame instance
}
\source{
This is a snapshot of all the terms available from 
GO.db (3.4.2), August 2017, using keys(GO.db, keytype="TERM").
}
\usage{
allGOterms
}
\description{
allGOterms: data.frame with ids and terms
}
\examples{
head(ontoProc::allGOterms)
}
\keyword{datasets}
