% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{getCellOnto}
\alias{getCellOnto}
\alias{getCellLineOnto}
\alias{getEFOOnto}
\alias{getChebiLite}
\alias{getCellosaurusOnto}
\alias{getUBERON_NE}
\title{load ontologies that may include non-ascii strings and therefore cannot be in data folder}
\usage{
getCellOnto()

getCellLineOnto()

getEFOOnto()

getChebiLite()

getCellosaurusOnto()

getUBERON_NE()
}
\value{
instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex
}
\description{
load ontologies that may include non-ascii strings and therefore cannot be in data folder
}
\examples{
co = getCellOnto()
co
clo = getCellLineOnto()
length(clo$id)
che = getChebiLite()
length(che$id)
efo = getEFOOnto()
length(efo$id)
}
