% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\docType{class}
\name{TermSet-class}
\alias{TermSet-class}
\alias{validObject}
\alias{show,TermSet-method}
\alias{show}
\title{manage ontological data with tags and a DataFrame instance}
\usage{
\S4method{show}{TermSet}(object)
}
\arguments{
\item{object}{instance of TermSet}
}
\value{
instance of TermSet
}
\description{
manage ontological data with tags and a DataFrame instance

validity method for TermSet

abbreviated display for TermSet instances
}
\examples{
validObject(new("TermSet"))
if (!exists(".efosupp")) .efosupp = buildEFOOntSupport()
defsibs = siblings_URL( model=getModel(.efosupp), 
               world=getWorld(.efosupp))
class(defsibs)
defsibs
}
