% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termProc.R
\docType{class}
\name{TermSet-class}
\alias{TermSet-class}
\alias{show,TermSet-method}
\alias{show}
\title{manage ontological data with tags and a DataFrame instance}
\usage{
\S4method{show}{TermSet}(object)
}
\arguments{
\item{object}{instance of TermSet class}
}
\value{
instance of TermSet
}
\description{
manage ontological data with tags and a DataFrame instance

abbreviated display for TermSet instances
}
\examples{
efoOnto = getEFOOnto()
defsibs = siblings_TAG("EFO:1001209", efoOnto)
class(defsibs)
defsibs
}
